Add additional coverage test (#29)
Test was reduced in order to cover the case where `cls` is true in
`search_check()` of `nlsat/nlsat_solver.cpp`. Additionally, it seems to
cover north of 600 additional lines, however, which are too many to
enumerate.

Reducing the test while maintaining coverage of the target line was
tricky, inlcuding things such as requiring `check_expr_same_sat()` to be
defined in a function, not being able to be inlined, or vector `fvs`
requiring a large number of elements be declared for some reason. A
cursory ASAN check seems to report an initial issue, which will be filed
separately shortly.

Co-authored-by: Andrei Lascu <andrei.lascu10@imperial.ac.uk>